backend/requirements.txt
bcrypt
Flask
Flask-Cors
Flask-Mail
google-auth
google-auth-oauthlib
itsdangerous
PyJWT
pymongo
python-dotenv
pytz
werkzeug
